requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
xgboost==2.0.3
plotly==5.19.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.27
fpdf==1.7.2
Pillow>=10.0.0
joblib>=1.4.0
scipy>=1.0.0
seaborn>=0.12.0
matplotlib>=3.7.0 